sql/calendar/requirements.txt
python-dateutil ~= 2.8.2
